apps/application/api_endpoint/Manufacturer/serializers.py
from rest_framework import serializers
from apps.application.models import Manufacturer



class ManufacturerCreateSerializer(serializers.ModelSerializer):
    class Meta:
        model = Manufacturer
        fields = "__all__"


class ManufacturerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Manufacturer
        fields = "__all__"