requirements.txt
# Django va asosiy framework
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.4.0
django-filter==23.5
django-redis==6.0.0
django-unfold==0.63.0
drf-spectacular==0.27.2

# Database
psycopg2-binary==2.9.10

# Telegram Bot
aiogram==3.21.0

# Celery va Redis
celery==5.5.3
redis==5.2.1

# HTTP va API
requests==2.32.4
aiohttp==3.12.15
aiofiles==24.1.0

# Environment va config
python-environ==0.4.54

# Image processing
# Pillow==10.2.0

# JWT va authentication
PyJWT==2.10.1

# Validation va serialization
pydantic==2.11.7
pydantic-core==2.33.2

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2

# Development va deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Async support
typing-extensions==4.14.1
typing-inspection==0.4.1

# Other dependencies
inflection==0.5.1
magic-filter==1.0.12
